2 small changes
just move the node operation in a new file node_engine &

substitute payment with check in line 40 of node_engine
